Clean up includes and move CPU voxelizer implementation methods to .cpp file
