+++ include/voxelized_geometry_tools/device_pointcloud_voxelization.hpp
@@ -1,6 +1,5 @@
 #pragma once
 
-#include <atomic>
 #include <cstdint>
 #include <map>
 #include <memory>
