+++ src/voxelized_geometry_tools/device_pointcloud_voxelization.cpp
@@ -5,10 +5,9 @@
 #endif
 
 #include <algorithm>
-#include <atomic>
+#include <chrono>
 #include <cmath>
 #include <cstdint>
-#include <fstream>
 #include <memory>
 #include <stdexcept>
 #include <vector>
